+++ Algebra/PartialOrd.hs
@@ -18,11 +18,10 @@ module Algebra.PartialOrd (
     gfpFrom, unsafeGfpFrom
   ) where
 
-import qualified Data.Set as S
+import qualified Data.IntMap as IM
 import qualified Data.IntSet as IS
 import qualified Data.Map as M
-import qualified Data.IntMap as IM
-
+import qualified Data.Set as S
 
 -- | A partial ordering on sets
 -- (<http://en.wikipedia.org/wiki/Partially_ordered_set>) is a set equipped